refactor: rename folder Space to space
