+++ source/dynamical_systems/include/dynamical_systems/Circular.hpp
@@ -8,9 +8,9 @@
 #include "dynamical_systems/DynamicalSystem.hpp"
 #include "state_representation/geometry/Ellipsoid.hpp"
 #include "state_representation/parameters/Parameter.hpp"
-#include "state_representation/Space/Cartesian/CartesianPose.hpp"
-#include "state_representation/Space/Cartesian/CartesianState.hpp"
-#include "state_representation/Space/Cartesian/CartesianTwist.hpp"
+#include "state_representation/space/Cartesian/CartesianPose.hpp"
+#include "state_representation/space/Cartesian/CartesianState.hpp"
+#include "state_representation/space/Cartesian/CartesianTwist.hpp"
 #include <cmath>
 #include <vector>
