+++ source/dynamical_systems/src/DynamicalSystem.cpp
@@ -1,8 +1,7 @@
 #include "dynamical_systems/DynamicalSystem.hpp"
 #include "state_representation/robot/JointState.hpp"
-#include "state_representation/Space/Cartesian/CartesianPose.hpp"
-#include "state_representation/Space/Cartesian/CartesianState.hpp"
-#include "state_representation/Space/Cartesian/CartesianTwist.hpp"
+#include "state_representation/space/Cartesian/CartesianPose.hpp"
+#include "state_representation/space/Cartesian/CartesianState.hpp"
 
 using namespace state_representation;
