+++ source/state_representation/include/state_representation/geometry/Ellipsoid.hpp
@@ -1,8 +1,3 @@
-/**
- * @author Baptiste Busch
- * @date 2019/04/16
- */
-
 #pragma once
 
 #include <list>
@@ -11,7 +6,7 @@
 #include <eigen3/Eigen/Sparse>
 #include <random>
 #include "state_representation/geometry/Shape.hpp"
-#include "state_representation/Space/Cartesian/CartesianPose.hpp"
+#include "state_representation/space/Cartesian/CartesianPose.hpp"
 
 namespace state_representation {
 /**
