+++ source/state_representation/include/state_representation/robot/Jacobian.hpp
@@ -8,8 +8,8 @@
 #include "state_representation/exceptions/IncompatibleSizeException.hpp"
 #include "state_representation/robot/JointTorques.hpp"
 #include "state_representation/robot/JointVelocities.hpp"
-#include "state_representation/Space/Cartesian/CartesianTwist.hpp"
-#include "state_representation/Space/Cartesian/CartesianWrench.hpp"
+#include "state_representation/space/Cartesian/CartesianTwist.hpp"
+#include "state_representation/space/Cartesian/CartesianWrench.hpp"
 #include "state_representation/State.hpp"
 #include <eigen3/Eigen/Core>
