+++ source/state_representation/include/state_representation/space/Cartesian/CartesianPose.hpp
@@ -5,8 +5,8 @@
 
 #pragma once
 
-#include "state_representation/Space/Cartesian/CartesianState.hpp"
-#include "state_representation/Space/Cartesian/CartesianTwist.hpp"
+#include "state_representation/space/Cartesian/CartesianState.hpp"
+#include "state_representation/space/Cartesian/CartesianTwist.hpp"
 
 namespace state_representation {
 class CartesianTwist;
