+++ source/state_representation/include/state_representation/space/Cartesian/CartesianState.hpp
@@ -6,7 +6,7 @@
 #pragma once
 
 #include "state_representation/exceptions/IncompatibleSizeException.hpp"
-#include "state_representation/Space/SpatialState.hpp"
+#include "state_representation/space/SpatialState.hpp"
 #include <eigen3/Eigen/Core>
 #include <eigen3/Eigen/Dense>
 #include <iostream>
