+++ source/state_representation/include/state_representation/space/Cartesian/CartesianWrench.hpp
@@ -5,7 +5,7 @@
 
 #pragma once
 
-#include "state_representation/Space/Cartesian/CartesianState.hpp"
+#include "state_representation/space/Cartesian/CartesianState.hpp"
 
 namespace state_representation {
 /**
