+++ source/state_representation/tests/testParameter.cpp
@@ -1,7 +1,7 @@
 #include "state_representation/parameters/Event.hpp"
 #include "state_representation/parameters/Parameter.hpp"
-#include "state_representation/Space/Cartesian/CartesianPose.hpp"
-#include "state_representation/Space/Cartesian/CartesianState.hpp"
+#include "state_representation/space/Cartesian/CartesianPose.hpp"
+#include "state_representation/space/Cartesian/CartesianState.hpp"
 #include <gtest/gtest.h>
 
 TEST(Conversion, PositiveNos) {
